+++ src/Compze.DocumentDb.MySql/Wiring/MySqlDocumentDbRegistrar.cs
@@ -5,6 +5,7 @@ using Compze.Internals.Sql.MySql;
 using Compze.Internals.Sql.MySql.Private;
 using Compze.Internals.Sql.MySql.Wiring;
 using Compze.TypeIdentifiers.Interning;
+using Compze.TypeIdentifiers.Interning.MySql.Wiring;
 using Layer = Compze.DocumentDb.MySql.MySqlDocumentDbSqlLayer;
 
 namespace Compze.DocumentDb.MySql.Wiring;
@@ -12,7 +13,8 @@ namespace Compze.DocumentDb.MySql.Wiring;
 public static class MySqlDocumentDbRegistrar
 {
    public static IComponentRegistrar MySqlDocumentDbSqlLayer(this IComponentRegistrar registrar) =>
-      registrar.MySqlSchemaContribution(Layer.SchemaCreationSql)
+      registrar.MySqlTypeIdInterner()
+               .MySqlSchemaContribution(Layer.SchemaCreationSql)
                .Register(
          Singleton.For<IDocumentDbSqlLayer>()
                   .CreatedBy((IMySqlConnectionPool connectionProvider, MySqlSqlLayerSchemaManager schemaManager, ITypeIdInterner typeIdInterner) => new Layer(connectionProvider, schemaManager, typeIdInterner)));