+++ src/Compze.DocumentDb.PostgreSql/Wiring/PgSqlDocumentDbRegistrar.cs
@@ -5,6 +5,7 @@ using Compze.Internals.Sql.PostgreSql;
 using Compze.Internals.Sql.PostgreSql.Private;
 using Compze.Internals.Sql.PostgreSql.Wiring;
 using Compze.TypeIdentifiers.Interning;
+using Compze.TypeIdentifiers.Interning.PostgreSql.Wiring;
 using Layer = Compze.DocumentDb.PostgreSql.PgSqlDocumentDbSqlLayer;
 
 namespace Compze.DocumentDb.PostgreSql.Wiring;
@@ -12,7 +13,8 @@ namespace Compze.DocumentDb.PostgreSql.Wiring;
 public static class PgSqlDocumentDbRegistrar
 {
    public static IComponentRegistrar PgSqlDocumentDbSqlLayer(this IComponentRegistrar registrar) =>
-      registrar.PgSqlSchemaContribution(Layer.SchemaCreationSql)
+      registrar.PgSqlTypeIdInterner()
+               .PgSqlSchemaContribution(Layer.SchemaCreationSql)
                .Register(
          Singleton.For<IDocumentDbSqlLayer>()
                   .CreatedBy((IPgSqlConnectionPool connectionProvider, PgSqlSqlLayerSchemaManager schemaManager, ITypeIdInterner typeIdInterner) => new Layer(connectionProvider, schemaManager, typeIdInterner)));