+++ src/Compze.Teventive.TeventStore.MicrosoftSql/Wiring/MsSqlTeventStoreRegistrar.cs
@@ -5,6 +5,7 @@ using Compze.Internals.Sql.MicrosoftSql;
 using Compze.Internals.Sql.MicrosoftSql.Private;
 using Compze.Internals.Sql.MicrosoftSql.Wiring;
 using Compze.TypeIdentifiers.Interning;
+using Compze.TypeIdentifiers.Interning.MicrosoftSql.Wiring;
 using Layer = Compze.Tessaging.Teventive.TeventStore.MicrosoftSql.MsSqlTeventStoreSqlLayer;
 
 namespace Compze.Tessaging.Teventive.TeventStore.MicrosoftSql.Wiring;
@@ -12,7 +13,8 @@ namespace Compze.Tessaging.Teventive.TeventStore.MicrosoftSql.Wiring;
 public static class MsSqlTeventStoreRegistrar
 {
    public static IComponentRegistrar MsSqlTeventStoreSqlLayer(this IComponentRegistrar registrar) =>
-      registrar.MsSqlSchemaContribution(Layer.SchemaCreationSql)
+      registrar.MsSqlTypeIdInterner()
+               .MsSqlSchemaContribution(Layer.SchemaCreationSql)
                .Register(
          Singleton.For<MsSqlTeventStoreConnectionManager>()
                   .CreatedBy((IMsSqlConnectionPool sqlConnectionProvider) => new MsSqlTeventStoreConnectionManager(sqlConnectionProvider)),