+++ src/Compze.Teventive.TeventStore.PostgreSql/Wiring/PgSqlTeventStoreRegistrar.cs
@@ -5,6 +5,7 @@ using Compze.Internals.Sql.PostgreSql;
 using Compze.Internals.Sql.PostgreSql.Private;
 using Compze.Internals.Sql.PostgreSql.Wiring;
 using Compze.TypeIdentifiers.Interning;
+using Compze.TypeIdentifiers.Interning.PostgreSql.Wiring;
 using Layer = Compze.Tessaging.Teventive.TeventStore.PostgreSql.PgSqlTeventStoreSqlLayer;
 
 namespace Compze.Tessaging.Teventive.TeventStore.PostgreSql.Wiring;
@@ -12,7 +13,8 @@ namespace Compze.Tessaging.Teventive.TeventStore.PostgreSql.Wiring;
 public static class PgSqlTeventStoreRegistrar
 {
    public static IComponentRegistrar PgSqlTeventStoreSqlLayer(this IComponentRegistrar registrar) =>
-      registrar.PgSqlSchemaContribution(Layer.SchemaCreationSql)
+      registrar.PgSqlTypeIdInterner()
+               .PgSqlSchemaContribution(Layer.SchemaCreationSql)
                .Register(
          Singleton.For<PgSqlTeventStoreConnectionManager>()
                   .CreatedBy((IPgSqlConnectionPool sqlConnectionProvider) => new PgSqlTeventStoreConnectionManager(sqlConnectionProvider)),